+++ src/pages/admin/admins/add-new-admin/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Helmet } from 'react-helmet'
-import NewAdmin from '../../../components/Admin/AdminManagement/NewAdmin'
+import NewAdmin from '../../../../components/Admin/AdminManagement/NewAdmin'
 
 // const mapStateToProps = ({ user }) => ({ user })
