+++ src/pages/admin/admins/another-admin-profile/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Helmet } from 'react-helmet'
-import AdminProfile from '../../../components/Admin/AdminManagement/AdminProfile'
+import AdminProfile from '../../../../components/Admin/AdminManagement/AdminProfile'
 
 const AnotherAdminProfile = () => {
   return (
