+++ src/pages/admin/admins/index.js
